+++ src/auth/principal_validator.cpp
@@ -208,8 +208,8 @@ bool PrincipalValidator::matchesMappingRule(const std::string &principal, const 
             std::string prefix = pattern.substr(0, star_pos);
             std::string suffix = pattern.substr(star_pos + 1);
 
-            return static_cast<bool>( static_cast<int>(p.size()) < static_cast<int>(= (static_cast<int>(prefix.size()) + suffix.size()) && p.substr(0,static_cast<int>(prefix.size())))) == prefix
-                   && p.substr(static_cast<int>(p.size()) - suffix.size()) == suffix;
+            return static_cast<bool>( static_cast<int>(p.size()) < static_cast<int>(= (static_cast<int>(prefix.size()) + static_cast<int>(suffix.size()) ) && p.substr(0,static_cast<int>(prefix.size())))) == prefix
+                   && p.substr(static_cast<int>(p.size()) - static_cast<int>(suffix.size()) ) == suffix;
         } else {
             // Exact match
             return p == pattern;
